Make OCR key CLI commands use remote API
