+++ core/cmd/local_client_test.go
@@ -670,162 +670,3 @@ func TestClient_SetNextNonce(t *testing.T) {
 	require.NotNil(t, key.NextNonce)
 	require.Equal(t, int64(42), *key.NextNonce)
 }
-
-func TestClient_P2P_CreateKey(t *testing.T) {
-	t.Parallel()
-	store, cleanup := cltest.NewStore(t)
-	defer cleanup()
-
-	app := new(mocks.Application)
-	app.On("GetStore").Return(store)
-
-	auth := cltest.CallbackAuthenticator{}
-	apiPrompt := &cltest.MockAPIInitializer{}
-	client := cmd.Client{
-		Config:                 store.Config,
-		AppFactory:             cltest.InstanceAppFactory{App: app},
-		KeyStoreAuthenticator:  auth,
-		FallbackAPIInitializer: apiPrompt,
-		Runner:                 cltest.EmptyRunner{},
-	}
-
-	set := flag.NewFlagSet("test", 0)
-	set.String("password", "../internal/fixtures/correct_password.txt", "")
-	c := cli.NewContext(nil, set, nil)
-
-	require.NoError(t, client.CreateP2PKey(c))
-
-	keys, err := app.GetStore().OCRKeyStore.FindEncryptedP2PKeys()
-	require.NoError(t, err)
-
-	// Created + fixture key
-	require.Len(t, keys, 2)
-
-	for _, e := range keys {
-		_, err = e.Decrypt(cltest.Password)
-		require.NoError(t, err)
-	}
-}
-
-func TestClient_P2P_DeleteKey(t *testing.T) {
-	t.Parallel()
-	store, cleanup := cltest.NewStore(t)
-	defer cleanup()
-
-	app := new(mocks.Application)
-	app.On("GetStore").Return(store)
-
-	auth := cltest.CallbackAuthenticator{}
-	apiPrompt := &cltest.MockAPIInitializer{}
-	client := cmd.Client{
-		Config:                 store.Config,
-		AppFactory:             cltest.InstanceAppFactory{App: app},
-		KeyStoreAuthenticator:  auth,
-		FallbackAPIInitializer: apiPrompt,
-		Runner:                 cltest.EmptyRunner{},
-	}
-
-	key, err := p2pkey.CreateKey()
-	require.NoError(t, err)
-	encKey, err := key.ToEncryptedP2PKey("password")
-	require.NoError(t, err)
-	err = store.OCRKeyStore.UpsertEncryptedP2PKey(&encKey)
-	require.NoError(t, err)
-
-	keys, err := store.OCRKeyStore.FindEncryptedP2PKeys()
-	require.NoError(t, err)
-	// Created  + fixture key
-	require.Len(t, keys, 2)
-
-	strID := strconv.FormatInt(int64(encKey.ID), 10)
-	set := flag.NewFlagSet("test", 0)
-	set.Parse([]string{strID})
-	c := cli.NewContext(nil, set, nil)
-
-	err = client.DeleteP2PKey(c)
-	require.NoError(t, err)
-
-	keys, err = store.OCRKeyStore.FindEncryptedP2PKeys()
-	require.NoError(t, err)
-	// fixture key only
-	require.Len(t, keys, 1)
-}
-
-func TestClient_CreateOCRKeyBundle(t *testing.T) {
-	t.Parallel()
-	store, cleanup := cltest.NewStore(t)
-	defer cleanup()
-
-	app := new(mocks.Application)
-	app.On("GetStore").Return(store)
-
-	auth := cltest.CallbackAuthenticator{}
-	apiPrompt := &cltest.MockAPIInitializer{}
-	client := cmd.Client{
-		Config:                 store.Config,
-		AppFactory:             cltest.InstanceAppFactory{App: app},
-		KeyStoreAuthenticator:  auth,
-		FallbackAPIInitializer: apiPrompt,
-		Runner:                 cltest.EmptyRunner{},
-	}
-
-	set := flag.NewFlagSet("test", 0)
-	set.String("password", "../internal/fixtures/correct_password.txt", "")
-	c := cli.NewContext(nil, set, nil)
-
-	require.NoError(t, client.CreateOCRKeyBundle(c))
-
-	keys, err := app.GetStore().OCRKeyStore.FindEncryptedOCRKeyBundles()
-	require.NoError(t, err)
-
-	// Created key + fixture key
-	require.Len(t, keys, 2)
-
-	for _, e := range keys {
-		_, err = e.Decrypt(cltest.Password)
-		require.NoError(t, err)
-	}
-}
-
-func TestClient_DeleteOCRKeyBundle(t *testing.T) {
-	t.Parallel()
-	store, cleanup := cltest.NewStore(t)
-	defer cleanup()
-
-	app := new(mocks.Application)
-	app.On("GetStore").Return(store)
-
-	auth := cltest.CallbackAuthenticator{}
-	apiPrompt := &cltest.MockAPIInitializer{}
-	client := cmd.Client{
-		Config:                 store.Config,
-		AppFactory:             cltest.InstanceAppFactory{App: app},
-		KeyStoreAuthenticator:  auth,
-		FallbackAPIInitializer: apiPrompt,
-		Runner:                 cltest.EmptyRunner{},
-	}
-
-	key, err := ocrkey.NewKeyBundle()
-	require.NoError(t, err)
-	encKey, err := key.Encrypt("password")
-	require.NoError(t, err)
-	err = store.OCRKeyStore.CreateEncryptedOCRKeyBundle(encKey)
-	require.NoError(t, err)
-
-	keys, err := store.OCRKeyStore.FindEncryptedOCRKeyBundles()
-	require.NoError(t, err)
-	// Created key + fixture key
-	require.Len(t, keys, 2)
-
-	set := flag.NewFlagSet("test", 0)
-	set.Parse([]string{key.ID.String()})
-	c := cli.NewContext(nil, set, nil)
-
-	err = client.DeleteOCRKeyBundle(c)
-	require.NoError(t, err)
-
-	keys, err = store.OCRKeyStore.FindEncryptedOCRKeyBundles()
-	require.NoError(t, err)
-	// Only fixture key remains
-	require.Len(t, keys, 1)
-}
