+++ core/cmd/remote_client.go
@@ -17,6 +17,8 @@ import (
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/chainlink/core/assets"
 	"github.com/smartcontractkit/chainlink/core/store/models"
+	"github.com/smartcontractkit/chainlink/core/store/models/ocrkey"
+	"github.com/smartcontractkit/chainlink/core/store/models/p2pkey"
 	"github.com/smartcontractkit/chainlink/core/store/presenters"
 	"github.com/smartcontractkit/chainlink/core/utils"
 	"github.com/smartcontractkit/chainlink/core/web"
@@ -756,4 +758,127 @@ func (cli *Client) CancelJobRun(c *clipkg.Context) error {
 		return cli.errorOut(errors.Wrap(err, "cli.parseResponse"))
 	}
 	return nil
+}
+
+func (cli *Client) CreateP2PKey(c *clipkg.Context) (err error) {
+	resp, err := cli.HTTP.Post("/v2/p2p_keys", nil)
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	defer func() {
+		if cerr := resp.Body.Close(); cerr != nil {
+			err = multierr.Append(err, cerr)
+		}
+	}()
+
+	if resp.StatusCode == 200 {
+		fmt.Println("Created P2P keypair.\n")
+	}
+	var key p2pkey.EncryptedP2PKey
+	return cli.renderAPIResponse(resp, &key)
+}
+
+func (cli *Client) ListP2PKeys(c *clipkg.Context) (err error) {
+	resp, err := cli.HTTP.Get("/v2/p2p_keys", nil)
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	defer func() {
+		if cerr := resp.Body.Close(); cerr != nil {
+			err = multierr.Append(err, cerr)
+		}
+	}()
+
+	var keys []p2pkey.EncryptedP2PKey
+	return cli.renderAPIResponse(resp, &keys)
+}
+
+func (cli *Client) DeleteP2PKey(c *clipkg.Context) (err error) {
+	if !c.Args().Present() {
+		return cli.errorOut(errors.New("Must pass the key ID to be deleted"))
+	}
+	id, err := strconv.ParseUint(c.Args().Get(0), 10, 32)
+	if err != nil {
+		return cli.errorOut(err)
+	}
+
+	resp, err := cli.HTTP.Delete(fmt.Sprintf("/v2/p2p_keys/%d", id))
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	defer func() {
+		if cerr := resp.Body.Close(); cerr != nil {
+			err = multierr.Append(err, cerr)
+		}
+	}()
+
+	if resp.StatusCode == 200 {
+		fmt.Println("P2P key deleted.\n")
+	}
+	var key p2pkey.EncryptedP2PKey
+	return cli.renderAPIResponse(resp, key)
+}
+
+// CreateOCRKeyBundle creates a key and inserts it into encrypted_ocr_key_bundles,
+// protected by the password in the password file
+func (cli *Client) CreateOCRKeyBundle(c *clipkg.Context) error {
+	resp, err := cli.HTTP.Post("/v2/off_chain_reporting_keys", nil)
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	defer func() {
+		if cerr := resp.Body.Close(); cerr != nil {
+			err = multierr.Append(err, cerr)
+		}
+	}()
+
+	if resp.StatusCode == 200 {
+		fmt.Println("Created OCR key bundle.\n")
+	}
+	var key ocrkey.EncryptedKeyBundle
+	return cli.renderAPIResponse(resp, &key)
+}
+
+// ListOCRKeyBundles lists the available OCR Key Bundles
+func (cli *Client) ListOCRKeyBundles(c *clipkg.Context) error {
+	resp, err := cli.HTTP.Get("/v2/off_chain_reporting_keys", nil)
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	defer func() {
+		if cerr := resp.Body.Close(); cerr != nil {
+			err = multierr.Append(err, cerr)
+		}
+	}()
+
+	var keys []ocrkey.EncryptedKeyBundle
+	return cli.renderAPIResponse(resp, &keys)
+}
+
+// DeleteOCRKeyBundle creates a key and inserts it into encrypted_ocr_keys,
+// protected by the password in the password file
+func (cli *Client) DeleteOCRKeyBundle(c *clipkg.Context) error {
+	if !c.Args().Present() {
+		return cli.errorOut(errors.New("Must pass the key ID to be deleted"))
+	}
+	id, err := models.Sha256HashFromHex(c.Args().Get(0))
+	if err != nil {
+		return cli.errorOut(err)
+	}
+
+	resp, err := cli.HTTP.Delete(fmt.Sprintf("/v2/off_chain_reporting_keys/%s", id))
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	defer func() {
+		if cerr := resp.Body.Close(); cerr != nil {
+			err = multierr.Append(err, cerr)
+		}
+	}()
+
+	if resp.StatusCode == 200 {
+		fmt.Println("OCR key bundle deleted.\n")
+	}
+	var key ocrkey.EncryptedKeyBundle
+	return cli.renderAPIResponse(resp, &key)
 }
